Templates
learned about the 'Pug', and via pug, I could make a base template that extedns to all the other html templates and partial templates that could be a part of a huge html template. Also learned about how to use conditionals in html(not sure if it is in html, since I only used it in pug environment).